services/interface/requirements.txt
gunicorn
flask-appconfig>=0.10
flask-bootstrap
flask-nav
flask-debug
flask-wtf
requests